1.rename class OperationContext -> InstanceOperationContext 2.move Map operation from NamingUtils to MapUtils
