+++ common/src/main/java/com/alibaba/nacos/common/utils/MapUtils.java
@@ -129,11 +129,11 @@ public class MapUtils {
     /**
      * ComputeIfAbsent lazy load.
      *
-     * @param target target Map data.
-     * @param key map key.
+     * @param target          target Map data.
+     * @param key             map key.
      * @param mappingFunction funtion which is need to be executed.
-     * @param param1 function's parameter value1.
-     * @param param2 function's parameter value1.
+     * @param param1          function's parameter value1.
+     * @param param2          function's parameter value1.
      * @return
      */
     @NotThreadSafe
@@ -154,4 +154,34 @@ public class MapUtils {
         return val;
     }
     
+    /**
+     * get target value from param, if not found will throw {@link IllegalArgumentException}.
+     *
+     * @param param param
+     * @param key   key
+     * @return value
+     */
+    public static String required(final Map<String, String> param, final String key) {
+        String value = param.get(key);
+        if (StringUtils.isEmpty(value)) {
+            throw new IllegalArgumentException("Param '" + key + "' is required.");
+        }
+        return value;
+    }
+    
+    /**
+     * get target value from param, if not found will return default value.
+     *
+     * @param param        param
+     * @param key          key
+     * @param defaultValue default value
+     * @return value
+     */
+    public static String optional(final Map<String, String> param, final String key, final String defaultValue) {
+        String value = param.get(key);
+        if (StringUtils.isBlank(value)) {
+            return defaultValue;
+        }
+        return value;
+    }
 }
